refactor(upgrader): clean up the UpgraderInterface and simplify the contract (#84)
