+++ contracts/upgrader/tests/utils/mod.rs
@@ -1,0 +1,2 @@
+mod dummy_contract;
+pub use dummy_contract::*;
